requirements.txt
affine==2.3.1
asttokens==2.0.8
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
easing-functions==1.0.4
entrypoints==0.3
executing==1.1.1
geographiclib==2.0
geopy==2.3.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.4.2
jupyter-core==4.11.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib-inline==0.1.6
more-itertools==4.2.0
nest-asyncio==1.5.6
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
packaging==21.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.31
protobuf==3.6.1
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2019.3
PyYAML==5.3.1
pyzmq==24.0.1
rasterio==1.3.4
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
snuggs==1.4.7
sos==4.1
ssh-import-id==5.10
stack-data==0.5.1
systemd-python==234
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
traitlets==5.4.0
Twisted==18.9.0
typing-extensions==4.1.1
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
zipp==1.0.0
zope.interface==4.7.1
